chore: intake url pattern updated
