+++ web/app/[workspaceSlug]/(projects)/projects/(detail)/[projectId]/inbox/loading.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { LogoSpinner } from "@/components/common";
+
+export default function InboxLoadingPage() {
+  return (
+    <div className="relative flex h-full w-full items-center justify-center">
+      <LogoSpinner />
+    </div>
+  );
+}
