+++ web/app/[workspaceSlug]/(projects)/projects/(detail)/[projectId]/intake/layout.tsx
